Rename Display\ImportAjax class
Renames to Import\Ajax.
